refactor: Split method invocation part out of FusionController
The EndpointInvoker still has knowledge of an HTTP request at this point. It will be refactored out later.
